+++ CnCTDRAMapEditor/Interface/ITechnoType.cs
@@ -28,6 +28,8 @@ namespace MobiusEditor.Interface
         bool IsFixedWing { get; }
         /// <summary>True if this object can harvest resources. This affects the default orders for placing it on the map.</summary>
         bool IsHarvester { get; }
+        /// <summary>True if this TechnoType is only available when enabling the game's expansion pack.</summary>
+        bool IsExpansionOnly { get; }
         /// <summary>True if this techno type adapts to its house colors.</summary>
         bool CanRemap { get; }
     }