+++ CnCTDRAMapEditor/Model/BuildingType.cs
@@ -102,9 +102,10 @@ namespace MobiusEditor.Model
         public string FactoryOverlay { get; private set; }
         public Bitmap Thumbnail { get; set; }
         public bool IsArmed => false; // Not actually true, but irrelevant for practical purposes; their Mission is not set in the ini file.
-        public bool IsHarvester => false;
         public bool IsAircraft => false;
         public bool IsFixedWing => false;
+        public bool IsHarvester => false;
+        public bool IsExpansionOnly => false;
 
         public bool IsFake => (this.Flag & BuildingTypeFlag.Fake) == BuildingTypeFlag.Fake;
         public bool HasTurret => (this.Flag & BuildingTypeFlag.Turret) == BuildingTypeFlag.Turret;